Addition of a logo!
Code
- Logo added
- Minor panel size changes to accomodate with a new design choice

After investigating the tree issue, it appears it is indeed a known limitiation.
Currently there are two choices, as I see it:
 - Keep the current tree, which is small
 - Attempt some wizardry with other resources to create a faux tree adhering to the designs

The former will be the current go-to unless a bump in the road occurs that would change the direction (unlikely)

That being said...
Post-commit development will stop for a few hours to do some research on 'what is to come' in tomorrows development, this would be:
 - Advanced menu!
 - Beginnings of web/interface interaction (maybe)

This push also signifies the end of the 'easy' part of development.