Big Update! Features - Travelling between frames now works correctly - New frame: Advanced - Working dark mode
Code
- Additional comments and reformating where appropriate
- More modularisation (in particular the logo)
- Listeners have been implemented (sort-of)

This update is fairly big. There may not be much to show, but a great deal of theory was learnt from this work.
Listeners have been investigated and implemented correctly!
Currently the task is to implement a sort-of 'global variable' to help track what configuration the programme is in (i.e. uStream or another service
This would massively aid modularity, of course. It would also make it seem like the developer working on it isn't a buffoon.

Next Update will likely contain
- Listener between frames (as described above)
- Settings file (need to think about this first)
- Maybe, just maybe, a start on the browsing area