Update: Note...
Rather than a typical update this commit is a note to say work was interrupted today and as such, there will be no commit today.
There is also a possibility there will be nothing tomorrow - although unlikely - as the developer has been absent due to a crisis.

Before work was interrupted today though, some development notes:

uStream
- The 'broadcaster list' is a list written in JavaScript
- Accessing this list is not possible unless a renderer is utilized
- Today, and much of the weekend, was spent attempting to figure out 'how to' get all of this working

As the above probably indiciates, efforts were unsuccessful, and because of this there will need to be a recorded delay on project work

Additional Note:
The work committed does not work and should not be looked at as a representation of the final product of development.