Update: Scraping is now functional!
After a series of trial and error-esque implementations, a functional script has now been found to extract all necessary data.

Refinement of this script and replication for other pages are now the goal, and after this commit, another should follow later today.

As an afterthought:
After the commit yesterday more research was gathered regarding 'how' scraping would occur, utilizing a 'headless' browser was the goal and found within selenium and an external driver. Currently there is a slight bit of concern with how this will pan out on release, as the developer is completely new to such things, but does realise the risk of the user needing the driver.

A further commit should be posted later tonight with additional layouts for tree selections, leading up to tomorrows big 'connect the data' day.

Although still behind, the project can still attain a reasonable finish date, it is believed.