Update: Enhanced script!
This enhanced script prints all four attributes needed into lists.
One list (broadcaster_viewers) needed to be cleared of whitespace, and code for this is now in place.

Additional 'ToDo' notes have been left in the code, as from this there are three other scripts to be written.

Next commit: Additional layouts for the tree!