Update: Thumbnails...
Many of the past work hours have been spent attempting to get thumbnails working, this has proven to be quite unsuccessful.

The problem:
To download the images there are two choices: hold them in a file or hold them in memory. Memory is seemingly impossible to due in conjunction with wxPython as of right now due to it's aversion of .jpeg image types. The file method hasn't been looked at too much but would mean some file parsing that may take too long to implement given the short time left for development overall;.

The solution:
A thumbnail that will represent all broadcaster thumbnails across the app. This is a dirty technique but it solves the issue.

Further update coming soon, hopefully implementing a 'page' scroll.