Update: Pages work!
Displaying the broadcast data is now possible. Six containers appear with unique information and two pages can be created after with unique data. All of this data is from the scraping method.

Currently there is an issue referring to the url:
With how the images are created and handled, there is no direct reference point outside of a top-layer index. This means whilst the objects are created and displayed uniquely the actions generated from these objects are not - namely the purpose of clicking on the thumbnail to then open a stream/open profile.

The next update intends to implement a solution to this (somehow), the alternative would be creating six unique images rather than six dynamic ones.

It has also been the decision that the log-in features will be cut for this cycle of development due to the complete 'locking' of the api by IBM, meaning there would be no meaningful development to that end.

When the aformentioned solution is implemented, the streamlink connection will begin to 'tie' things together for the demo. This SHOULD all occur tomorrow (next commit).

After this will come the implementation of two (not one, but two as per an advisement) 'search' bars.