Update: Streams now open!
Referring back to the previous update: each button now has a unique identifier (the URL) that can be accessed on click.

When the user clicks the thumbnail, the URL is passed into the command line directly and ordered to open the URL at a set quality. This then opens the stream in the default player, at aforementioned quality.

More about this update:

Initially it was the plan to integrate PyGObject into the programme - similarly to how the official streamlink documentation shows it - as it wasn't REALLY the plan to directly use the command line in reference to streamlink, however after several hours of troubleshooting attempting to JUST INSTALL PyGObject this has been scrapped for the 'dirtier' approach which is quite a margin easier in terms of total code but now needs another requirement: a configuration file.

The configuration file for Streamlink in Windows isnt' explicit, so utilising our configuration file for this purpose would be the quickest solution. This would allow:
- Saving of preferred 'quality'
- Saving of preferred 'player'
- Potentially more, if time is plentiful (somewhat likely)

Next update should be the big 'search bar' update, followed by the 'how do I package this for a demo, oh god why' update.