Update: Minor Bug Fixing
Not much to commit today, as the developer has decided to take the day off in favor of working the rest of the week.

Code
- Bug fix: advanced not opening
- Addition of second search bar

The bug fix was in relation the previous update that bound a 'browse' event to the sidebar tree. This overwrit the 'advanced' event, and once realized, this mistake was easily fixed by creating a singular event that can be added to in the future.

Next update: 'streamlink' search bar + .exe wrapping 101