Update: Profiles! Selection refresh! Bug investigation!
New
- Profiles
 - Default layout
 - Activated by clicking title on broadcaster
 - Uniquely generated

- Selection refresh now works appropriately

On Thursday a demo was presented to in another environment. Everything went well with the exception of a web socket error that caused the application to freeze. Some cursory research leads the developer to believe this would be a difficult to fix problem as:
- The developer has no knowledge on web socket errors
- The error appears to be because of the environemtns network settings
- Replicating the error would mean travelling and working off-site

This problem will be looked into further on Monday or future integration testing

For the remainder of todays allocated work hours, there will be an investigation into why only 15 items can be accessed in the lists, yet clearly the lists hold more. Other odd bugs/errors will also be looked at before the big 'replicate scripts and get it all working' day.

Monday will possibly be one of the last days of development for this cycle and will see a mop-up of all ToDo's and 'bonus' features.