Update: Everything works!
Today is a fairly big commit. All options on the side tree are now responsive, including all options that use the web scraper.

New
- All browse options functional
- All log-in options functional (as a place-holder)
- All option options functional (two were removed, will condense into one tomorrow)
- Section for 'Upcoming'
- Update on profile section
- Profile is updated from the webscraper, and completely functional!

Work today looked at two different things along side the above
- Limit of '15' on a list
- Search bar implementation

The 'limit of 15' issue stems from the web scraper it would seem, as different pages tend to return differing amounts of information. The maximum any can return without throwing up an index error is currently 14. There are no plans to fix this as the error seems to lie entirely with the data returned from uStream, and any further investigation would likely lead to little progress, especially due to the small amount of time left.

Today the search bar was supposed to be implemented. Everything was prepared, researched and ready...
When going to uStream to get the HTML needed, there was no search bar present.
The developer was - confused - but there appears to be some HTML references left from it, so the question is now "where did it go?" rather than "was it ever there?" (to which the developer can attest to it being there. If this returns wthin the next 48 hours, the developer will look into implementing it, if not...an unfortunate cut.

Tomorrow will see more config file addtions to the application in an attempt to help modularise it, alongside the search bar if it re-appears.