Update: 1.0 Release
New
- Full 20 results are now available for EVERY page
- Quality settings are available under 'Advanced'
- Thumbnails are downloaded AND displayed perfectly
- RuntimeError on the sidebar tree "fixed"

While the developer was in transit a thought occurred, "could it be, the viewers are being returned impartially, not the entire list?". This was so. As it turns out, the viewers were the source of the iffy numbers being returned, and a minor fix has enabled full functionality for every page.

Quality settings have been implemented in a basic form, adhering to 'worst' and 'best' as of now.

Thumbnails are downloaded locally, stored in a file and given the index names, then assigned to the thumbnail locally using the index name and a bitmap convert. They work fully.

Minor runtime error that occurred when the application switched between 'gui' and 'service select' has been caught and patched over.

As of now, the application is in a "1.0" state. Everything, in terms of available features, have been implemented. There are currently three tasks left to do:
- Implement search bar (can't as of now; not available on uStream)
- Modularise application further (would be minor but not fully necessary)
- Improve layout/style (this needs to be done, but not right away)

At this point - testing begins.